fix: apply valid Cypher escaping everywhere + prove #2409 paths on Windows CI
Follow-ups from #2425 review, addressed in-PR:

1. LadybugDB rejects SQL-style '' quote doubling in Cypher string
   literals (parser error, not an escape). Beyond the writeback sites
   fixed in the previous commit, the same broken pattern lived in the
   augment engine (pattern word, filePath, nodeId batch list), the
   embedding semantic-search metadata lookup (nodeId batch list), and
   five wiki graph queries (file lists, process ids) — every one
   producing a query that never parsed, silently, wherever the call
   site swallowed errors. All nine now use a shared escapeCypherString
   (src/core/lbug/cypher-escape.ts, backslash-escaping, backslashes
   before quotes), which lbug-adapter also imports. The csv-generator
   '' usages are CSV-reader grammar, not Cypher — deliberately
   untouched, documented on the helper.

2. Windows CI coverage for the #2409 fixes. The failing environment is
   Windows and this devcontainer is Linux/WSL2, so the platform-
   sensitive pieces are routed through the windows-latest cross-
   platform matrix instead: lbug-delete-nodes-for-files (chunked
   IN-list DETACH DELETE + escaping against the real native addon)
   joins LBUG_NATIVE, and the dirty-recovery parking test is split out
   of the orchestration suite into incremental-dirty-recovery.test.ts
   (sidecar renames next to a live native DB — exactly the semantics
   that differ on Windows) so the matrix runs it without paying for the
   whole 2.5-minute suite. sidecar-recovery.test.ts was already in the
   matrix, so the parking unit tests run on Windows as-is.

Local verification: cypher-escape 4/4, dirty-recovery split 1/1,
trimmed orchestration suite 10/10, adapter/conn/delete/sidecar 71/71,
augmentation + embedding-pipeline + wiki suites 64/64; tsc and
prettier clean.
