+++ gitnexus/src/core/wiki/graph-queries.ts
@@ -6,6 +6,7 @@
  */
 
 import { initLbug, executeQuery, closeLbug, touchRepo, pinRepo } from '../lbug/pool-adapter.js';
+import { escapeCypherString } from '../lbug/cypher-escape.js';
 
 const REPO_ID = '__wiki__';
 
@@ -146,7 +147,7 @@ export async function getInterFileCallEdges(): Promise<CallEdge[]> {
 export async function getIntraModuleCallEdges(filePaths: string[]): Promise<CallEdge[]> {
   if (filePaths.length === 0) return [];
 
-  const fileList = filePaths.map((f) => `'${f.replace(/'/g, "''")}'`).join(', ');
+  const fileList = filePaths.map((f) => `'${escapeCypherString(f)}'`).join(', ');
   const rows = await executeQuery(
     REPO_ID,
     `
@@ -174,7 +175,7 @@ export async function getInterModuleCallEdges(filePaths: string[]): Promise<{
 }> {
   if (filePaths.length === 0) return { outgoing: [], incoming: [] };
 
-  const fileList = filePaths.map((f) => `'${f.replace(/'/g, "''")}'`).join(', ');
+  const fileList = filePaths.map((f) => `'${escapeCypherString(f)}'`).join(', ');
 
   const outRows = await executeQuery(
     REPO_ID,
@@ -221,7 +222,7 @@ export async function getInterModuleCallEdges(filePaths: string[]): Promise<{
 export async function getProcessesForFiles(filePaths: string[], limit = 5): Promise<ProcessInfo[]> {
   if (filePaths.length === 0) return [];
 
-  const fileList = filePaths.map((f) => `'${f.replace(/'/g, "''")}'`).join(', ');
+  const fileList = filePaths.map((f) => `'${escapeCypherString(f)}'`).join(', ');
 
   // Find processes that have steps in the given files
   const procRows = await executeQuery(
@@ -247,7 +248,7 @@ export async function getProcessesForFiles(filePaths: string[], limit = 5): Prom
     const stepRows = await executeQuery(
       REPO_ID,
       `
-      MATCH (s)-[r:CodeRelation {type: 'STEP_IN_PROCESS'}]->(p:Process {id: '${procId.replace(/'/g, "''")}'})
+      MATCH (s)-[r:CodeRelation {type: 'STEP_IN_PROCESS'}]->(p:Process {id: '${escapeCypherString(procId)}'})
       RETURN s.name AS name, s.filePath AS filePath, labels(s)[0] AS type, r.step AS step
       ORDER BY r.step
     `,
@@ -295,7 +296,7 @@ export async function getAllProcesses(limit = 20): Promise<ProcessInfo[]> {
     const stepRows = await executeQuery(
       REPO_ID,
       `
-      MATCH (s)-[r:CodeRelation {type: 'STEP_IN_PROCESS'}]->(p:Process {id: '${procId.replace(/'/g, "''")}'})
+      MATCH (s)-[r:CodeRelation {type: 'STEP_IN_PROCESS'}]->(p:Process {id: '${escapeCypherString(procId)}'})
       RETURN s.name AS name, s.filePath AS filePath, labels(s)[0] AS type, r.step AS step
       ORDER BY r.step
     `,
